+++ spotinst_sdk/__init__.py
@@ -9,6 +9,11 @@ import argparse
 
 from spotinst_sdk import aws_elastigroup
 from spotinst_sdk import spotinst_functions
+from spotinst_sdk import spotinst_emr
+from spotinst_sdk import spotinst_stateful
+from spotinst_sdk import spotinst_blue_green_deployment
+from spotinst_sdk import spotinst_deployment_action
+from spotinst_sdk import spotinst_asg
 
 VAR_SPOTINST_SHARED_CREDENTIALS_FILE = 'SPOTINST_SHARED_CREDENTIALS_FILE'
 VAR_SPOTINST_PROFILE = 'SPOTINST_PROFILE'
@@ -30,8 +35,14 @@ _SpotinstClient__spotinst_sdk_user_agent = '{}/{}'.format(
 
 class SpotinstClient:
     __account_id_key = "accountId"
+    __base_url = "https://api.spotinst.io/aws/ec2"
     __base_elastigroup_url = "https://api.spotinst.io/aws/ec2/group"
+    __base_emr_url = "https://api.spotinst.io/aws/emr/mrScaler"
     __base_functions_url = "https://api.spotinst.io/functions"
+    __base_stateful_url = "https://api.spotinst.io/aws/ec2/statefulMigrationGroup"
+    __base_kube_url = "https://api.spotinst.io/mcs/kubernetes/cluster"
+    __base_saving_url = "https://api.spotinst.io/aws/potentialSavings"
+
     camel_pat = re.compile(r'([A-Z])')
     under_pat = re.compile(r'_([a-z])')
 
@@ -72,11 +83,51 @@ class SpotinstClient:
         else:
             self.set_log_level(log_level)
 
+
+    # region EMR
+    def create_emr(self, emr):
+        emr = spotinst_emr.EMRCreationRequest(emr)
+
+        excluded_group_dict = self.exclude_missing(json.loads(emr.toJSON()))
+        
+        formatted_group_dict = self.convert_json(
+            excluded_group_dict, self.underscore_to_camel)
+
+        body_json = json.dumps(formatted_group_dict)
+        
+        group_response = self.send_post(
+            body=body_json,
+            url=self.__base_emr_url,
+            entity_name='emr')
+
+        formatted_response = self.convert_json(
+            group_response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["items"][0]
+
+        return retVal
+
     # endregion
 
-    # region Elastigroup
-    def create_elastigroup(self, group):
+    # region Kubernetes
 
+    def get_kubernetes_cluster_cost(self, custer_id, from_date, to_date):        
+        geturl = self.__base_kube_url + "/" + custer_id + "/costs"
+        query_params = self.build_query_params_with_input({"toDate":to_date, "fromDate":from_date})
+
+        result = self.send_get(url=geturl, query_params=query_params, entity_name='kubernetes')
+
+        formatted_response = self.convert_json(
+            result, self.camel_to_underscore)
+
+        return formatted_response["response"]["items"][0]
+
+    # endregion
+
+
+
+    # region Elastigroup
+    def create_elastigroup(self, group):        
         group = aws_elastigroup.ElastigroupCreationRequest(group)
 
         excluded_group_dict = self.exclude_missing(json.loads(group.toJSON()))
@@ -89,8 +140,8 @@ class SpotinstClient:
         self.print_output(body_json)
 
         group_response = self.send_post(
-            body_json,
-            self.__base_elastigroup_url,
+            body=body_json,
+            url=self.__base_elastigroup_url,
             entity_name='elastigroup')
 
         formatted_response = self.convert_json(
@@ -145,11 +196,11 @@ class SpotinstClient:
         self.print_output(body_json)
 
         group_response = self.send_put(
-            body_json,
             self.__base_elastigroup_url +
             "/" +
             group_id,
-            entity_name='elastigroup')
+            entity_name='elastigroup',
+            body=body_json)
 
         formatted_response = self.convert_json(
             group_response, self.camel_to_underscore)
@@ -209,10 +260,25 @@ class SpotinstClient:
             content, self.camel_to_underscore)
         return formatted_response["response"]["items"]
 
-    def roll_group(self, group_id, group_roll):
+    def get_elastigroup_activity(self, group_id, start_date):
+        query_params = self.build_query_params_with_input({"fromDate":start_date})
 
+        content = self.send_get(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/events",
+            query_params=query_params,
+            entity_name='active events')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+        return formatted_response["response"]["items"]
+
+    def roll_group(self, group_id, group_roll):
         group_roll_request = aws_elastigroup.ElastigroupRollRequest(
             group_roll=group_roll)
+
 
         excluded_group_roll_dict = self.exclude_missing(
             json.loads(group_roll_request.toJSON()))
@@ -221,8 +287,6 @@ class SpotinstClient:
             excluded_group_roll_dict, self.underscore_to_camel)
 
         body_json = json.dumps(formatted_group_roll_dict)
-
-        self.print_output(body_json)
 
         roll_response = self.send_put(
             url=self.__base_elastigroup_url +
@@ -235,9 +299,359 @@ class SpotinstClient:
         formatted_response = self.convert_json(
             roll_response, self.camel_to_underscore)
 
+        retVal = formatted_response["response"]
+
+        return retVal
+
+    def get_all_group_deployment(self, group_id):
+        content = self.send_get(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/roll",
+            entity_name='roll')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+        return formatted_response["response"]["items"]
+
+
+    def get_deployment_status(self, group_id, roll_id):
+        content = self.send_get(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/roll/"+
+                str(roll_id),
+            entity_name='roll')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+
+        return formatted_response["response"]["items"]
+
+    def stop_deployment(self, group_id, roll_id):
+        content = self.send_put(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/roll/"+
+                str(roll_id),
+            body=json.dumps(dict(roll=dict(status="STOPPED"))),
+            entity_name='roll')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+
+        return formatted_response["response"]
+
+    def create_deployment_action(self, group_id, roll_id, deployment_action):
+        deployment_action_request = spotinst_deployment_action.DeploymentActionRequest(deployment_action)
+
+        deployment_action_dict = self.exclude_missing(
+            json.loads(deployment_action_request.toJSON()))
+
+        formatted_group = self.convert_json(
+            deployment_action_dict, self.underscore_to_camel)
+
+        body_json = json.dumps(formatted_group)
+
+        detach_response = self.send_post(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/roll/"+
+                str(roll_id),
+            body=body_json,
+            entity_name='roll')
+
+        formatted_response = self.convert_json(
+            detach_response, self.camel_to_underscore)
+
         retVal = formatted_response["response"]["items"]
 
         return retVal
+
+    def get_instance_type_by_region(self, region):
+        query_params = dict(region=region)
+        response = self.send_get(
+            url=self.__base_url+
+            "/spotType",
+            query_params=query_params,
+            entity_name="instance"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"]
+
+    def lock_instance(self, instance_id, lock_time=None):
+        query_params= dict(ttlInMinutes=lock_time)
+
+        response = self.send_post(
+            url=self.__base_url +
+            "/instance/" +
+            instance_id +
+            "/lock",
+            query_params=query_params,
+            entity_name="instance"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["status"]
+
+    def unlock_instance(self, instance_id):
+        response = self.send_post(
+            url=self.__base_url +
+            "/instance/" +
+            instance_id +
+            "/unlock",
+            entity_name="instance"
+        )
+        
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["status"]
+
+    def enter_instance_standby(self, instance_id):
+        response = self.send_post(
+            url=self.__base_url + 
+            "/instance/" + 
+            instance_id +
+            "/standby/enter",
+            entity_name="instance"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["status"]
+
+    def exit_instance_standby(self, instance_id):
+        response = self.send_post(
+            url=self.__base_url + 
+            "/instance/" + 
+            instance_id +
+            "/standby/exit",
+            entity_name="instance"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["status"]
+
+    def get_instance_status(self, instance_id):
+        response = self.send_get(
+            url=self.__base_url + 
+            "/instance/" +
+            instance_id,
+            entity_name="instance"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"][0]
+
+    def get_instance_healthiness(self, group_id):
+        response = self.send_get(
+            url=self.__base_elastigroup_url +
+             "/" + group_id + 
+            "/instanceHealthiness",
+            entity_name="instance"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"][0]  
+
+
+    def create_instance_signal(self, instance_id, signal):
+        body = dict(instanceId=instance_id, signal=signal)
+
+        response = self.send_post(
+            url= self.__base_url + 
+            "/instance/signal",
+            body=body,
+            entity_name="instance"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["status"]  
+
+    def get_cost_per_account(self, to_date=None, from_date=None):
+        query_params=dict(toDate=to_date, fromDate=from_date)
+
+        response = self.send_get(
+            url="https://api.spotinst.io/aws/costs",
+            query_params=query_params,
+            entity_name="cost"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"]  
+
+    def get_cost_per_elastigroup(self, group_id, to_date=None, from_date=None):
+        query_params=dict(toDate=to_date, fromDate=from_date)
+
+        response = self.send_get(
+            url=self.__base_elastigroup_url +
+            "/" + group_id +
+            "/costs",
+            query_params=query_params,
+            entity_name="cost"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"]   
+
+    def get_group_detailed_cost(self, group_id, to_date=None, from_date=None):
+        query_params=dict(toDate=to_date, fromDate=from_date)
+
+        response = self.send_get(
+            url=self.__base_elastigroup_url +
+            "/" + group_id +
+            "/costs/detailed",
+            query_params=query_params,
+            entity_name="cost"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"] 
+
+    def get_potential_savings(self):
+        response = self.send_get(
+            url="https://api.spotinst.io/aws/potentialSavings",
+            entity_name="saving"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"] 
+
+    def get_instance_potential_savings(self, instance_ids, region):
+        instance_str = ""
+
+        for instance in instance_ids:
+            instance_str += instance + "," 
+        
+        query_params = dict(region=region, instanceIds=instance_str)
+
+        response = self.send_get(
+            url="https://api.spotinst.io/aws/instancePotentialSavings",
+            query_params=query_params,
+            entity_name="saving"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"] 
+
+    def list_suspended_scaling_policies(self, group_id):
+        response = self.send_get(
+            url=self.__base_elastigroup_url +
+            "/" + group_id + 
+            "/scale/suspensions",
+            entity_name="scaling policies"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"]   
+
+    def suspend_scaling_policies(self, group_id, policy_name):
+        query_params = dict(policyName=policy_name)
+
+        response = self.send_post(
+            url=self.__base_elastigroup_url +
+            "/" + group_id + 
+            "/scale/suspendPolicy",
+            query_params=query_params,
+            entity_name="scaling policies"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"][0]
+
+    def resume_suspended_scaling_policies(self, group_id, policy_name):
+        query_params = dict(policyName=policy_name)
+
+        response = self.send_post(
+            url=self.__base_elastigroup_url +
+            "/" + group_id + 
+            "/scale/resumePolicy",
+            query_params=query_params,
+            entity_name="scaling policies"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["status"]
+
+    def list_suspended_process(self, group_id):
+        response = self.send_get(
+            url=self.__base_elastigroup_url +
+            "/" + group_id + 
+            "/suspension",
+            entity_name="suspend process"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"]
+
+    def suspend_process(self, group_id, processes, suspensions):
+        body = dict(suspensions=suspensions, processes=processes)
+        
+        response = self.send_post(
+            url=self.__base_elastigroup_url +
+            "/" + group_id + 
+            "/suspension",
+            body=body,
+            entity_name="suspend process"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response["response"]["items"]
+
+    def remove_suspended_process(self, group_id, processes):
+        body = dict(processes=processes)
+
+        response = self.send_delete(
+            url=self.__base_elastigroup_url +
+            "/" + group_id + 
+            "/suspension",
+            body=body,
+            entity_name="suspend process"
+        )
+
+        formatted_response = self.convert_json(
+            response, self. camel_to_underscore)
+
+        return formatted_response
 
     def detach_elastigroup_instances(self, group_id, detach_configuration):
 
@@ -252,8 +666,6 @@ class SpotinstClient:
 
         body_json = json.dumps(formatted_group_detach_dict)
 
-        self.print_output(body_json)
-
         detach_response = self.send_put(
             url=self.__base_elastigroup_url +
                 "/" +
@@ -266,6 +678,319 @@ class SpotinstClient:
             detach_response, self.camel_to_underscore)
 
         retVal = formatted_response["response"]["status"]
+
+        return retVal
+
+    def import_stateful_instance(self, stateful_instance):
+        stateful_instance = spotinst_stateful.StatefulImportRequest(stateful_instance)
+
+        excluded_group_dict = self.exclude_missing(json.loads(stateful_instance.toJSON()))
+
+        formatted_group_dict = self.convert_json(
+            excluded_group_dict, self.underscore_to_camel)
+
+        body_json = json.dumps(formatted_group_dict)
+        
+        group_response = self.send_post(
+            body=body_json,
+            url=self.__base_stateful_url,
+            entity_name='import stateful instance')
+
+        formatted_response = self.convert_json(
+            group_response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["items"][0]
+
+        return retVal
+
+    def get_stateful_import_status(self, stateful_migration_id):
+        content = self.send_get(
+            url=self.__base_stateful_url +
+                "/" +
+                str(stateful_migration_id),
+            entity_name='get stateful import status')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+
+        return formatted_response["response"]["items"] 
+
+    def delete_stateful_import(self, stateful_migration_id):
+        content = self.send_delete(
+            url=self.__base_stateful_url +
+                "/" +
+                str(stateful_migration_id),
+            entity_name='delete stateful import')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+        return formatted_response
+
+    def deallocate_stateful_instance(self, group_id, stateful_instance_id):
+        content = self.send_put(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/statefulInstance/" +
+                str(stateful_instance_id +
+                "/deallocate"),
+            entity_name='deallocate stateful instance')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+        return formatted_response["response"]
+
+    def recycle_stateful_instance(self, group_id, stateful_instance_id):
+        content = self.send_put(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/statefulInstance/" +
+                str(stateful_instance_id +
+                "/recycle"),
+            entity_name='recycle stateful instance')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+        return formatted_response["response"]
+
+
+    def get_stateful_instances(self, group_id):
+        content = self.send_get(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/statefulInstance",
+            entity_name='get stateful instance')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+        return formatted_response["response"]["items"]           
+
+    def resume_stateful_instance(self, group_id, stateful_instance_id):
+        content = self.send_put(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/statefulInstance/" +
+                str(stateful_instance_id +
+                "/resume"),
+            entity_name='resume stateful instance')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+        return formatted_response["response"]
+
+    def pause_stateful_instance(self, group_id, stateful_instance_id):
+        content = self.send_put(
+            url=self.__base_elastigroup_url +
+                "/" +
+                str(group_id) +
+                "/statefulInstance/" +
+                str(stateful_instance_id) +
+                "/pause",
+            entity_name='pause stateful instance')
+
+        formatted_response = self.convert_json(
+            content, self.camel_to_underscore)
+
+        return formatted_response["response"]
+
+
+
+    def beanstalk_maintenance_status(self, group_id):
+        
+        status_response = self.send_get(
+            url=self.__base_elastigroup_url+
+            "/" +
+            str(group_id) + 
+            "/beanstalk/maintenance/status",
+            entity_name="beanstalk maintenance start")
+
+        formatted_response = self.convert_json(
+            status_response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["status"]
+
+        return retVal
+
+
+    def beanstalk_maintenance_start(self, group_id):
+
+        start_response = self.send_put(
+            url=self.__base_elastigroup_url+
+            "/" +
+            str(group_id) + 
+            "/beanstalk/maintenance/start",
+            body={},
+            entity_name="beanstalk maintenance start")
+
+        formatted_response = self.convert_json(
+            start_response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["status"]
+
+        return retVal
+
+    def beanstalk_maintenance_finish(self, group_id):   
+
+        finish_response = self.send_put(
+            url=self.__base_elastigroup_url+
+            "/" +
+            str(group_id) + 
+            "/beanstalk/maintenance/finish",
+            body={},
+            entity_name="beanstalk maintenance start")
+
+        formatted_response = self.convert_json(
+            finish_response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["status"]
+
+        return retVal
+
+    def beanstalk_import(self, region, env_id=None, env_name=None):
+        query_params = dict(region=region, environmentId=env_id, environmentName=env_name)
+
+        response = self.send_get(
+            url=self.__base_elastigroup_url +
+            "/beanstalk/import",
+            query_params=query_params,
+            entity_name="beanstalk import"
+        )
+
+        formatted_response = self.convert_json(
+            response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["items"][0]
+
+        return retVal
+
+    def beanstalk_reimport(self, group_id):
+        response = self.send_put(
+            url=self.__base_elastigroup_url +
+            "/" + str(group_id) + 
+            "/beanstalk/reimport",
+            entity_name="beanstalk reimport"
+        )
+
+        formatted_response = self.convert_json(
+            response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["items"][0]
+
+        return retVal
+
+    def import_asg(self, region, asg_name, asg, dry_run=None):
+        query_params = dict(region=region, autoScalingGroupName=asg_name, dryRun=dry_run)
+
+        asg = spotinst_asg.ImportASGRequest(asg)
+
+        excluded_group_dict = self.exclude_missing(json.loads(asg.toJSON()))
+
+        formatted_group_dict = self.convert_json(
+            excluded_group_dict, self.underscore_to_camel)
+
+        body_json = json.dumps(formatted_group_dict)
+        
+        response = self.send_post(
+            body=body_json,
+            url=self.__base_elastigroup_url+
+            "/autoScalingGroup/import",
+            query_params=query_params,
+            entity_name='import asg')
+
+        print(json.dumps(response))
+
+        formatted_response = self.convert_json(
+            response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["items"][0]
+
+        return retVal
+
+    def get_activity_events(self, group_id, from_date):
+        query_params = dict(fromDate=from_date)
+
+        response = self.send_get(
+            url=self.__base_elastigroup_url +
+            "/" + group_id + "/events",
+            query_params=query_params,
+            entity_name="activity groups"
+        )
+
+        formatted_response = self.convert_json(
+            response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["items"]
+
+        return retVal
+
+    def ami_backup(self, group_id):
+        response = self.send_post(
+            url=self.__base_elastigroup_url +
+            "/" + group_id + "/amiBackup",
+            entity_name="ami backup"
+        )
+
+        print(json.dumps(response))
+
+        formatted_response = self.convert_json(
+            response, self.camel_to_underscore)
+
+        retVal = formatted_response
+
+        return retVal["response"]["status"]
+
+    def create_blue_green_deployment(self, group_id, blue_green_deployment):
+        blue_green_deployment = spotinst_blue_green_deployment.BlueGreenDeploymentRequest(blue_green_deployment)
+
+        excluded_group_dict = self.exclude_missing(json.loads(blue_green_deployment.toJSON()))
+
+        formatted_group_dict = self.convert_json(
+            excluded_group_dict, self.underscore_to_camel)
+
+        body_json = json.dumps(formatted_group_dict)
+
+        group_response = self.send_post(
+            body=body_json,
+            url=self.__base_elastigroup_url + "/" + group_id + "/codeDeploy/blueGreenDeployment",
+            entity_name='create b/g deployment')
+
+        print(group_response)
+
+        formatted_response = self.convert_json(
+            group_response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["items"][0]
+
+        return retVal
+
+    def get_blue_green_deployment(self, group_id):
+        response = self.send_get(
+            url= self.__base_elastigroup_url + "/" + group_id + "/codeDeploy/blueGreenDeployment",
+            entity_name="get b/g deployment")
+
+        formatted_response = self.convert_json(
+            response, self.camel_to_underscore)
+
+        retVal = formatted_response["response"]["items"][0]
+
+        return retVal
+
+    def stop_blue_green_deployment(self, group_id, deployment_id):
+        response = self.send_put(
+            url=self.__base_elastigroup_url + "/" + group_id + "/codeDeploy/blueGreenDeployment/" + deployment_id + "/stop",
+            entity_name="stop b/g deployment")
+
+        print(response)
+
+        formatted_response = self.convert_json(
+            response, self.camel_to_underscore)
+
+        print(formatted_response)
+
+        retVal = formatted_response["response"]
 
         return retVal
 
@@ -286,8 +1011,8 @@ class SpotinstClient:
         self.print_output(body_json)
 
         app_response = self.send_post(
-            body_json,
-            self.__base_functions_url +
+            body=body_json,
+            url=self.__base_functions_url +
             '/application',
             entity_name='application')
 
@@ -312,8 +1037,8 @@ class SpotinstClient:
         self.print_output(body_json)
 
         env_response = self.send_post(
-            body_json,
-            self.__base_functions_url +
+            body=body_json,
+            url=self.__base_functions_url +
             '/environment',
             entity_name='environment')
 
@@ -338,8 +1063,8 @@ class SpotinstClient:
         self.print_output(json.dumps(formatted_fx_dict))
 
         fx_response = self.send_post(
-            body_json,
-            self.__base_functions_url +
+            body=body_json,
+            url=self.__base_functions_url +
             '/function',
             entity_name='function')
 
@@ -353,9 +1078,13 @@ class SpotinstClient:
     # endregion
 
     # region Utils
-    def send_get(self, url, entity_name):
+    def print_output(self, output):
+        if self.should_print_output is True:
+            print(output)
+
+    def send_get(self, url,entity_name,query_params=None):
         agent = self.resolve_user_agent()
-        query_params = self.build_query_params()
+        query_params = query_params or self.build_query_params()
         headers = dict(
             {
                 'User-Agent': agent,
@@ -365,18 +1094,16 @@ class SpotinstClient:
         )
 
         self.print_output("Sending get request to spotinst API.")
-
         result = requests.get(url, params=query_params, headers=headers)
 
         if result.status_code == requests.codes.ok:
             self.print_output("Success")
-
-            data = json.loads(result.content)
+            data = json.loads(result.content.decode('utf-8'))
             return data
         else:
             self.handle_exception("getting {}".format(entity_name), result)
 
-    def send_delete(self, url, entity_name):
+    def send_delete(self, url, entity_name, body=None):
         agent = self.resolve_user_agent()
         query_params = self.build_query_params()
         headers = dict(
@@ -389,7 +1116,7 @@ class SpotinstClient:
 
         self.print_output("Sending deletion request to spotinst API.")
 
-        result = requests.delete(url, params=query_params, headers=headers)
+        result = requests.delete(url, params=query_params, body=body, headers=headers)
 
         if result.status_code == requests.codes.ok:
             self.print_output("Success")
@@ -422,9 +1149,9 @@ class SpotinstClient:
         else:
             self.handle_exception("deleting {}".format(entity_name), result)
 
-    def send_post(self, body, url, entity_name):
+    def send_post(self, url, entity_name, body=None, query_params=None):
         agent = self.resolve_user_agent()
-        query_params = self.build_query_params()
+        query_params = query_params or self.build_query_params()
         headers = dict(
             {
                 'User-Agent': agent,
@@ -443,12 +1170,12 @@ class SpotinstClient:
 
         if result.status_code == requests.codes.ok:
             self.print_output("Success")
-            data = json.loads(result.content)
+            data = json.loads(result.content.decode('utf-8'))
             return data
         else:
             self.handle_exception("creating {}".format(entity_name), result)
 
-    def send_put(self, body, url, entity_name):
+    def send_put(self, url, entity_name, body=None):
         agent = self.resolve_user_agent()
         query_params = self.build_query_params()
         headers = dict(
@@ -468,7 +1195,7 @@ class SpotinstClient:
 
         if result.status_code == requests.codes.ok:
             self.print_output("Success")
-            data = json.loads(result.content)
+            data = json.loads(result.content.decode('utf-8'))
             return data
         else:
             self.handle_exception("updating {}".format(entity_name), result)
@@ -495,7 +1222,7 @@ class SpotinstClient:
 
         if result.status_code == requests.codes.ok:
             self.print_output("Success")
-            data = json.loads(result.content)
+            data = json.loads(result.content.decode('utf-8'))
             return data
         else:
             self.handle_exception("updating {}".format(entity_name), result)
@@ -510,7 +1237,8 @@ class SpotinstClient:
     def handle_exception(self, action_string, result):
         self.print_output(result.status_code)
 
-        data = json.loads(result.content)
+        data = json.loads(result.content.decode('utf-8'))
+
         response_json = json.dumps(data["response"])
         self.print_output(response_json)
 
@@ -525,6 +1253,7 @@ class SpotinstClient:
             return val
         elif type(val) in (int, float, bool, "".__class__, u"".__class__):
             return val
+
         for k, v in list(val.items()):
             new_v = v
             if isinstance(v, dict):